Better keybindings: hold C while dragging to select the crop region
The Alt key (in combination with clicking and dragging) is used by various window managers
to move windows around and to resize them. Even third party tools (like AltDrag on Windows)
tend to use this sort of keybinding. Ergo: it's a bad idea to bind Alt+Drag. Instead, we're
now holding C (for crop) while dragging to select the crop region.

Alongside this change, I've also moved the Alt-Hold shortcut for showing just the reference
image in comparison mode to Shift-Hold (which is more consistent with other reference-related
shortcuts, too).