SWITCH TO GPLv3 LICENSE
As I'm adding more features to tev, more and more dependencies become
more ergonomic to use if tev is under the GPLv3 license. Over the years,
I've also become more and more appreciative of the values that
free-software licenses represent, so this is a good opportunity to make
the switch. tev's version will get bumped to 2.0 to reflect this major
change in licensing.