feat: add CLI flag to disable loading of gainmaps
Also overhauls loading of images: instead of checking for format
compatibility in a separate step from loading, both are now handled in
the same (ImageLoader::load) function. A special type of exception is
instead thrown if the format is deemed to be unsupported, in which case
the next loader in will be tried in order. This doesn't just simplify a
whole bunch of logic but also avoids loading image data into memory
twice when loading Ultra HDR JPEG images.