feat(wayland): configurable TEV_CM_SDR_WHITE_LEVEL envvar
This envvar allows configuring tev's brightness in mixed HDR/SDR
compositing situations. on wayland This is only needed if the compositor
does not treat SDR sRGB values of 1 as 80 nits (e.g. when having a
custom brightness configures on Hyprland or KDE). On Windows, the
compositor technically also does not treat the SDR white level as 80
nits, but the actual white level is exposed at the OS level and read by
tev already.