feat(channels): asynchronous texture upload
If the backend supports it, schedule an async copy that uses DMA to
copy the texture without blocking the host. The operation is part of
the graphics queue and correctly ordered wrt. other display
operations. On Apple M* GPUs, CPU/GPU share the same memory, so this
step just converts the texture into a more suitable layout.

Co-authored-by: Thomas Müller <git@tom94.net>