fix(imageviewer): label default tonemap as "None" instead of "sRGB"
Clarifies that tev's default tone mapping displays images with the
colors they were encoded with. While tev *technically* converts the
image's colors to sRGB space at some point in the pipeline when this
option is set, this is only one step that ultimately leads the image to
appear to screen in a different (potentially wide or HDR) color space
that has nothing to do with sRGB.

Ergo: change the label to avoid confusing users.